pyproject.toml
[project]
name = "edutrack-engine"
version = "0.1.0"
description = "Curriculum Intelligence Engine - Fetches, validates, and generates educational content from official curricula"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "EduTrack Team"}
]

dependencies = [
    "pydantic>=2.0.0",
    "google-generativeai>=0.8.0",
    "langgraph>=0.2.0",
    "redis>=5.0.0",
    "supabase>=2.0.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
